backend/requirements.txt
Flask
pywinrm
requests-credssp
flask-cors
